AI-Plat/platform/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# 数据处理
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# 本体处理
rdflib==7.0.0
owlready2==0.40
sparqlwrapper==2.0.0

# 图数据库
neo4j==5.14.1
py2neo==2023.0

# 搜索引擎
elasticsearch==8.11.0
elastic-transport==8.11.0

# 大模型接口
openai==1.3.7
langchain==0.0.336
transformers==4.35.2
torch==2.1.1

# 消息队列
kafka-python==2.0.2
redis==5.0.1

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# 容器与部署
docker==7.1.0
kubernetes==28.1.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2

# 其他工具
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0